Fixes up sprite map rendering to be seamless.

+++ include/apsis/sync/clock.h
@@ -16,7 +16,7 @@ namespace Apsis {
       /*
        *  Returns: The fraction of seconds since the last call to elapsedTime.
        */
-      double elapsedTime();
+      float elapsedTime();
 
     private:
       // Internal storage; depends on platform
